hybrid-array: add slice_as_chunks(_mut) methods, minor refactoring
